requirements.txt
streamlit==1.32.0
langchain==0.1.4
langchain-community==0.0.16
pypdf==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-generativeai==0.3.2
faiss-cpu
numpy